requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-instrumentation-sqlalchemy==0.41b0
opentelemetry-instrumentation-logging==0.41b0
sqlalchemy==1.4.46  
clickhouse-sqlalchemy==0.2.7
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0